requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.11.4
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==7.2.1
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langsmith==0.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.27.1
numpy==1.26.4
openai==1.63.2
orjson==3.10.15
packaging==23.2
pandas==2.2.3
pillow==10.4.0
pinecone==6.0.1
pinecone-plugin-interface==0.0.7
postgrest==0.19.3
propcache==0.3.0
protobuf==4.25.6
pyarrow==19.0.1
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pypdf==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
realtime==2.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
storage3==0.11.3
streamlit==1.31.0
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tenacity==8.5.0
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
validators==0.34.0
watchdog==6.0.0
websockets==14.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
